Remove unused stuff from the code
Attempt to begin cleanup and review of code.  Removed unused pin
access, and unused functions.  Also change commented out code for
printing debug info back to the way that was previously implemented:
checking the state of the "printing" variable.  More cleanup to go,
but this code should still function the same.